try to make features compose properly
