+++ raphtory/Cargo.toml
@@ -81,11 +81,11 @@ tokio = { version = "1.27.0", features = ["full"]} # for vector testing
 dotenv = "0.15.0" # for vector testing
 
 [features]
-default = ["search", "vectors"]
+default = []
 # Enables the graph loader io module
 io = ["dep:zip", "dep:neo4rs", "dep:bzip2", "dep:flate2", "dep:csv", "dep:serde_json", "dep:reqwest", "dep:tokio"]
 # Enables generating the pyo3 python bindings
-python = ["io", "vectors", "dep:pyo3", "dep:pyo3-asyncio", "dep:num", "dep:display-error-chain", "dep:arrow2"]
+python = ["io", "dep:pyo3", "dep:pyo3-asyncio", "dep:num", "dep:display-error-chain", "dep:arrow2"]
 # search
 search = ["dep:tantivy"]
 # vectors
