+++ raphtory/src/python/graph/mod.rs
@@ -2,6 +2,8 @@ pub mod algorithm_result;
 pub mod edge;
 pub mod graph;
 pub mod graph_with_deletions;
+
+#[cfg(feature = "search")]
 pub mod index;
 pub mod node;
 pub mod pandas;
