+++ raphtory/src/python/packages/mod.rs
@@ -1,4 +1,6 @@
 pub mod algorithms;
 pub mod graph_gen;
 pub mod graph_loader;
+
+#[cfg(feature = "vectors")]
 pub mod vectors;
